version: dynamically create file with git commit hash
Dynamically generate a file with the current git commit hash at build
time.

The question remains how this will be handled if we ever release the
software, because the release tarball will not contain this file. Then
again, the release tarball will also not include Libopencm3.

Releases should include a prebuilt tarball with the binary and ELF
outputs.